Лабы/Лаб6/ansible/playbooks/app/requirements.txt
Flask
mysql-connector